+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 
 ### Fix
 
+* Exiftool plugin will scan the path variable for executable [CTC-1585]
 * Robot Freezes when hitting cancel on auto-save dialog [CTC-1636]
 * RegexServiceImpl.getMatcher uses incorrect timeout [CTC-1629]
 * Web.download now creates all required folders [CTC-1532]